% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.R
\name{annot_from_gr}
\alias{annot_from_gr}
\title{Annotate a RangedSummarizedExperiment using Granges objects}
\usage{
annot_from_gr(obj, gr, cols_to_map)
}
\arguments{
\item{obj}{RangedSummarizedExperiment}

\item{gr}{GRanges with annotations to map to obj}

\item{cols_to_map}{list of columns from gr to map
to obj. If the list has names, the names will be used for
colnames in the obj}
}
\description{
Utility function to map annotations from GRanges
to rowRanges of SummarizedExperiment. If multiple features overlap
then they will be concatenated as comma separated values.
}
