% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bam.R
\name{bam_to_df}
\alias{bam_to_df}
\title{Read in bam file as a data.frame}
\usage{
bam_to_df(filename = NULL, region = ".", tags = NULL)
}
\arguments{
\item{filename}{path to bam file}

\item{region}{samtools region query string (i.e. chr1:100-1000)}

\item{tags}{bam tags to return, supplied with type suffix, i.e. "XS:A" for character,
"AS:i" for integer, and "CB:Z" for string. Bam entries with missing tags will return empty
strings.}
}
\description{
Read in bam file as a data.frame
}
\examples{
bam_file <- system.file("extdata", "small_sorted.bam", package = "kentr")
res <- bam_to_df(bam_file)
head(res)

res <- bam_to_df(bam_file,
                 region = "chr1:3e6-3.1e6",
                 tags = c("XS:A", "AS:i", "CB:Z"))
res
}
