% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BedFile-class}
\alias{BedFile-class}
\alias{.BedFile}
\title{A reference class for generating and storing index of BedFile}
\description{
Class to store information about a bedfile used
for indicating regions for pileup
}
\section{Fields}{

\describe{
\item{\code{.extptr}}{an externalptr to c-level samtools bedindex}

\item{\code{path}}{filepath to bed file}

\item{\code{open}}{logical indicating if the index is open}
}}


