% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-utils.R
\name{build_tag_index}
\alias{build_tag_index}
\title{Build index for tag sorted bam file}
\usage{
build_tag_index(bamfn)
}
\arguments{
\item{bamfn}{CB tag sorted bamfile}
}
\value{
name of index generated, which is the bam file + ".bri"
}
\description{
Build index for tag sorted bam file
}
