% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BedFile.R
\name{close.BedFile}
\alias{close.BedFile}
\title{Close a BedFile index connection}
\usage{
\method{close}{BedFile}(con, ...)
}
\arguments{
\item{con}{BedFile class}

\item{...}{present for consistency}
}
\description{
Close a BedFile index connection
}
