% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_editing.R
\name{count_edits}
\alias{count_edits}
\title{Counts edits}
\usage{
count_edits(
  se,
  edit_frequency = 0.01,
  min_count = 10,
  edit_from = NULL,
  edit_to = NULL
)
}
\arguments{
\item{se}{A SummarizedExperiment object created by \code{create_se} and
processed by \code{calc_edit_frequency}}

\item{edit_frequency}{OPTIONAL the edit frequency used to determine the number of sites.
Default is 0.01.}

\item{min_count}{OPTIONAL the number of reads used to determine the number of edited sites.
Default is 10.}

\item{edit_from}{OPTIONAL if not using a pre-built type, you can specify
your own editing. This should be a nucleotide (A, C, G, or T) and should
correspond to the nucleotide you expect in the reference. Ex. for A to I
editing events, this would be "A". If type is not "AI", both edit from
and edit_to must be set.}

\item{edit_to}{OPTIONAL if not using a pre-built type, you can specify
your own editing. This should be a nucleotide (A, C, G, or T) and should
correspond to the nucleotide you expect after the editing event. Ex. for A to I
editing events, this would be "G". If type is not "AI", both edit from
and edit_to must be set.}
}
\description{
Counts edits per sample and add new colData columns with the number of edited sites
(n_sites) and the  fraction of edits (edit_idx). This function should be called by
\code{calc_edit_frequency} and is not meant to be used directly.
}
