% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{download_GSE99249}
\alias{download_GSE99249}
\title{Download GSE99249 BAM files and related data}
\usage{
download_GSE99249(verbose = TRUE)
}
\arguments{
\item{verbose}{report messages}
}
\value{
A named list with paths to BAM files, a FASTA file and a bed
file of known editing sites from hg38 chromosome 18.
}
\description{
This function will download ~ 250 MB of data and store files in a
BiocFileCache.
}
\examples{
\donttest{
download_GSE99249()
}

}
\seealso{
Other external-data: 
\code{\link{download_NA12878}()},
\code{\link{download_human_pbmc}()}
}
\concept{external-data}
