% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{download_human_pbmc}
\alias{download_human_pbmc}
\title{Download 10x PMBC bam file and related data}
\usage{
download_human_pbmc(verbose = TRUE)
}
\arguments{
\item{verbose}{report messages}
}
\value{
A named list with paths to bam file, fasta file,
bed file of editing_sites, and an .rds file with a
SingleCellExperiment
}
\description{
This function will download < 1 GB of data and store files in a
BiocFileCache.
}
\examples{
\donttest{
download_human_pbmc()
}

}
\seealso{
Other external-data: 
\code{\link{download_GSE99249}()},
\code{\link{download_NA12878}()}
}
\concept{external-data}
