% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{download_NA12878}
\alias{download_NA12878}
\title{Download NA12878 BAM files and related data}
\usage{
download_NA12878(path)
}
\arguments{
\item{path}{path to directory to download data}
}
\value{
A named list with paths to an RNA-seq and WGS BAM file, and a FASTA file
from hg38 chromosome 4.
}
\description{
This function will download ~ 5 GB of data.
}
\examples{
\dontrun{
td <- tempdir()
download_NA12878(td)
}

}
\seealso{
Other external-data: 
\code{\link{download_GSE99249}()},
\code{\link{download_human_pbmc}()}
}
\concept{external-data}
