% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{filter_multiallelic}
\alias{filter_multiallelic}
\title{Filter out multi-allelic sites}
\usage{
filter_multiallelic(se)
}
\arguments{
\item{se}{\code{SummarizedExperiment::SummarizedExperiment}}
}
\description{
Remove sites with multiple variant bases from a
\code{SummarizedExperiment}. \code{rowData()} gains a new column, \code{Var}, that
contains the variant allele detected at each site.
}
\examples{
example(create_se, echo = FALSE)
filter_multiallelic(se)

}
\seealso{
Other se-filters: 
\code{\link{filter_clustered_variants}()},
\code{\link{filter_splice_variants}()}
}
\concept{se-filters}
