% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{get_splice_sites}
\alias{get_splice_sites}
\title{Extract regions surrounding splice sites}
\usage{
get_splice_sites(txdb, slop = 4)
}
\arguments{
\item{txdb}{A TxDb object}

\item{slop}{The number of bases upstream and downstream of splice site to extract}
}
\description{
This function will return intervals containing splice sites and
adjacent regions.
}
\examples{
if(require(TxDb.Hsapiens.UCSC.hg38.knownGene)){
  txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
  res <- get_splice_sites(txdb)
  res[1:5]
}
}
