% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{find_mispriming_sites}
\alias{find_mispriming_sites}
\title{Find regions with oligodT mispriming}
\usage{
find_mispriming_sites(
  bamfile,
  fafile,
  pos_5p = 5,
  pos_3p = 20,
  min_reads = 2,
  tag = "pa",
  tag_values = 6:300,
  n_reads_per_chunk = 1e+06,
  verbose = TRUE
)
}
\description{
OligodT will prime at A-rich regions in an RNA. Reverse transcription
from these internal priming sites will install an oligodT sequence at the 3' end
of the cDNA. Sequence variants within these internal priming sites are enriched
for variants converting the genomic sequence to the A encoded by the oligodT primer.
Trimming poly(A) from the 3' ends of reads reduces but does not eliminate these signals

This function will identify regions that are enriched for mispriming events. Reads
that were trimmed to remove poly(A) (encoded in the pa tag by 10x genomics) are
identified. The aligned 3' positions of these reads are counted, and sites passing
thresholds (at least 2 reads) are retained as possible sites of mispriming. Be default
regions 5 bases upstream and 20 bases downstream of these putative mispriming sites
are returned.
}
\examples{
bam_fn <- raer_example("5k_neuron_mouse_possort.bam")
fa_fn <- raer_example("mouse_tiny.fasta")
find_mispriming_sites(bam_fn, fa_fn)

}
