% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pileup.R
\name{get_pileup}
\alias{get_pileup}
\title{Generate pileup}
\usage{
get_pileup(
  bamfile = NULL,
  fafile = NULL,
  bedfile = NULL,
  region = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL,
  outfile = "tmp.txt",
  return_data = TRUE
)
}
\arguments{
\item{bamfile}{path to bam file}

\item{fafile}{path to fasta file}

\item{bedfile}{path to bed file with sites to query}

\item{region}{samtools region query string (i.e. chr1:100-1000)}

\item{chrom}{chrom to process, not to be used with region}

\item{start}{start position, not to be used with region}

\item{end}{end position , not to be used with region}

\item{outfile}{output tabix index file}

\item{return_data}{return data as a Granges table?}
}
\description{
Generate pileup
}
