% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCoverageAtPositions}
\alias{getCoverageAtPositions}
\title{Copy of VariantTools::extractCoverageForPositions
Authors: Michael Lawrence, Jeremiah Degenhardt, Robert Gentleman}
\usage{
getCoverageAtPositions(cov, pos)
}
\arguments{
\item{cov}{coverage produced by \code{\link[GenomicAlignments:coverage-methods]{GenomicAlignments::coverage()}}}

\item{pos}{GRanges containing editing sites}
}
\description{
Copy of VariantTools::extractCoverageForPositions
Authors: Michael Lawrence, Jeremiah Degenhardt, Robert Gentleman
}
