% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-BedFile.R, R/methods-RegionIndex.R
\docType{class}
\name{BedFile-class}
\alias{BedFile-class}
\alias{.BedFile}
\alias{RegionIndex-class}
\alias{.RegionIndex}
\alias{indexBed}
\alias{indexRegions}
\alias{indexBed,character-method}
\alias{close.BedFile}
\alias{indexRegions,GRanges-method}
\alias{close.RegionIndex}
\alias{print.RegionIndex}
\title{A reference class for generating and storing index of BedFile}
\usage{
indexBed(file, ...)

indexRegions(gr, ...)

\S4method{indexBed}{character}(file)

\method{close}{BedFile}(con, ...)

\S4method{indexRegions}{GRanges}(gr)

\method{close}{RegionIndex}(con, ...)

\method{print}{RegionIndex}(con, ...)
}
\arguments{
\item{file}{path to bed file}

\item{...}{present for consistency}

\item{gr}{GRanges object}

\item{con}{RegionIndex class}
}
\value{
Return an object of class BedFile which contains a pointer
to the bed file index.

Return an object of class RegionIndex which contains a pointer
to the region index.
}
\description{
Class to store information about a bedfile used
for indicating regions for pileup

Class to store information about regions used
for indicating regions for pileup
}
\section{Fields}{

\describe{
\item{\code{.extptr}}{an externalptr to c-level samtools bedindex}

\item{\code{path}}{filepath to bed file}

\item{\code{open}}{logical indicating if the index is open}

\item{\code{.extptr}}{an externalptr to c-level samtools regidx}

\item{\code{gr}}{GRanges containing regions}

\item{\code{open}}{logical indicating if the index is open}
}}


\examples{
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
indexBed(bed_fn)
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
bed <- indexBed(bed_fn)
close(bed)
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
gr <- import(bed_fn)
indexRegions(gr)
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
bed <- indexBed(bed_fn)
close(bed)
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
gr <- import(bed_fn)
i <- indexRegions(gr)
print(i)
}
