% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-RegionIndex.R
\name{indexRegions,GRanges-method}
\alias{indexRegions,GRanges-method}
\alias{close.RegionIndex}
\alias{print.RegionIndex}
\title{Generate a samtools style index for regions}
\usage{
\S4method{indexRegions}{GRanges}(gr)

\method{close}{RegionIndex}(con, ...)

\method{print}{RegionIndex}(con, ...)
}
\arguments{
\item{gr}{GRanges object}

\item{con}{RegionIndex class}

\item{...}{present for consistency}
}
\value{
Return an object of class RegionIndex which contains a pointer
to the region index.
}
\description{
Generate a samtools style index for regions

Close a RegionIndex index connection

Print a RegionIndex index connection
}
\examples{
library(rtracklayer)
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
gr <- import(bed_fn)
gr <- tile(gr, width = 1) |> unlist()
strand(gr) <- sample(c("+", "-"), length(gr), replace = TRUE)
gr$ref <- sample(c("A", "T", "C", "G"), length(gr), replace = TRUE)
gr$alt <- sample(c("A", "T", "C", "G"), length(gr), replace = TRUE)
indexRegions(gr)
library(rtracklayer)
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
bed <- indexBed(bed_fn)
close(bed)
library(rtracklayer)
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
gr <- import(bed_fn)
i <- indexRegions(gr)
print(i)
}
