% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BedFile.R
\name{indexBed,character-method}
\alias{indexBed,character-method}
\title{Generate a samtools style index for a BedFile}
\usage{
\S4method{indexBed}{character}(file)
}
\arguments{
\item{file}{path to bed file}
}
\value{
Return an object of class BedFile which contains a pointer
to the bed file index.
}
\description{
Generate a samtools style index for a BedFile
}
