% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{make_editing_plots}
\alias{make_editing_plots}
\title{Makes summary plots of editing}
\usage{
make_editing_plots(
  se_object,
  colors = NULL,
  meta_col = "genotype_treatment",
  replicate = "rep"
)
}
\arguments{
\item{se_object}{A SummarizedExperiment object created by \code{create_se}}

\item{colors}{OPTIONAL The colors of the replicates. If no colors are
provided, Set1 from `RColorBrewer will be used}

\item{meta_col}{The column in colData to be used to separate out samples
based on the condition. For example, "genotype", "treatment", or
"genotype_treatment." Default is "genotype_treatment."}

\item{replicate}{The column in colData contining information about the
replicates. Default is "rep".}
}
\description{
Generates plots of the number of sites edited per sample and the percent
of editing events per sample. This function is written to be called directly
by \code{add_editing_frequencies}
}
