% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pileups.R
\name{merge_pileups}
\alias{merge_pileups}
\title{Merge pileups into a RangedSummarizedExperiment}
\usage{
merge_pileups(
  plps,
  assay_cols = c("ALT", "nRef", "nAlt", "nA", "nT", "nC", "nG"),
  sample_names = NULL,
  sparse = FALSE,
  fill_na = NULL,
  verbose = FALSE
)
}
\arguments{
\item{plps}{results from running \code{\link[=pileup_sites]{pileup_sites()}}, can be one result, a list of
results, or a named list of results. If a named list is given, the colData
will be named using the names in the list.}

\item{assay_cols}{character vector of columns to store as assays}

\item{sample_names}{A list of names to be added to the SE object. If no
sample names are given and plps is not a named list, then default names (ie
sample_1, sample_2, ..., sample_n) will be given and a warning will be
printed.}

\item{sparse}{If TRUE, numeric matrices will be stored as sparseMatrices. All
missing values will be coerced to 0 in the sparseMatrices.}

\item{fill_na}{Numeric value to replace NAs in numeric matrices Should only
be used when plps were computed independently with a min_nucleotide_count =
1, otherwise sites may be set to 0, although they may have coverage > 0 but
less than the min_nucleotide_count parameter. Not applied when sparse =
TRUE, which coerces missing values to 0.}

\item{verbose}{print information on progress}
}
\value{
\code{RangedSummarizedExperiment} populated with assays specified in
\code{assay_cols}.
}
\description{
Create a \code{RangedSummarizedExperiment} from a single or list of
pileups (e.g., for different samples) generated by \code{pileup_sites()}.
}
\keyword{internal}
