% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mock_rse}
\alias{mock_rse}
\title{Generate a small RangedSummarizedExperiment object for tests and examples}
\source{
\url{https://www.ncbi.nlm.nih.gov/bioproject/PRJNA386593}
}
\usage{
mock_rse()
}
\value{
RangedSummarizedExperiment populated with pileup data
}
\description{
A RangedSummarizedExperiment containing a subset of data from an RNA-seq experiment
to measure the effects of IFN treatment of cell lines with wild-type
or ADAR1-KO.
}
\examples{
mock_rse()

}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/29395325/}
}
