% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bam.R
\name{pileup}
\alias{pileup}
\title{Generate pileup}
\usage{
pileup(filename = NULL, region = ".")
}
\arguments{
\item{filename}{path to bam file}

\item{region}{samtools region query string (i.e. chr1:100-1000)}
}
\description{
Generate pileup
}
