% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raer.R
\docType{package}
\name{raer}
\alias{raer}
\alias{raer-package}
\title{raer: RNA editing tools in R}
\description{
Toolkit for identification, visualization, and statistical testing of RNA editing signals from within R.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/rnabioco/raer}
  \item Report bugs at \url{https://github.com/rnabioco/raer/issues}
}

}
\author{
\strong{Maintainer}: Kent Riemondy \email{kent.riemondy@cuanschutz.edu} (\href{https://orcid.org/0000-0003-0750-1273}{ORCID})

Authors:
\itemize{
  \item Kristen Wells \email{kristen.wells-wrasman@cuanschutz.edu} (\href{https://orcid.org/0000-0002-7466-8164}{ORCID})
  \item Ryan Sheridan \email{ryan.sheridan@cuanschutz.edu}
}

Other contributors:
\itemize{
  \item RNA Bioscience Initiative [funder]
}

}
