% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raer_example}
\alias{raer_example}
\title{Provide working directory for raer example files.}
\usage{
raer_example(path)
}
\arguments{
\item{path}{path to file}
}
\description{
Provide working directory for raer example files.
}
\examples{
raer_example('human.fasta')

}
