% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup.R
\name{read_pileup}
\alias{read_pileup}
\title{Read pileup, indexed by tabix}
\usage{
read_pileup(tbx_fn, region = NULL)
}
\arguments{
\item{tbx_fn}{filename}

\item{region}{region to read from file, samtools style
region specifiers are supported.}
}
\value{
\code{GenomicRanges::GRanges} containing positions of editing sites and pileup
data.
}
\description{
This function can read in pileup or site data tables produced by
pileup_sites().
}
\examples{
bamfn <- system.file("extdata", "SRR5564269_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")
plp_fn <- tempfile()
plp <- pileup_sites(bamfn, fafn, return_data = FALSE, outfile_prefix = plp_fn)

# first table contains site information, second table pileup count
lapply(plp, function(x) head(read_pileup(x)))
unlink(c(plp, plp_fn, paste0(plp, ".tbi")))

}
