% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-pileup.R
\name{read_sparray}
\alias{read_sparray}
\title{Read tables produced by pileup_cells()}
\usage{
read_sparray(mtx_fn, sites_fn, bc_fn)
}
\arguments{
\item{mtx_fn}{.mtx.gz file path}

\item{sites_fn}{sites.txt.gz file path}

\item{bc_fn}{bcs.txt.gz file path}
}
\value{
a list of \code{sparseMatrix}, of \code{NULL} if mtx_fn is empty
}
\description{
Read in tables produced by \code{pileup_cells()}, which are an
extension of the matrixMarket sparse matrix format to store values for more
than 1 matrix.

The .mtx.gz files are formatted with columns:
\enumerate{
\item row index (0 based)
\item column index (0 based)
\item values for sparseMatrix #1 (nRef)
\item values for sparseMatrix #2 (nVar)
N) values for sparseMatrix ... (...) ununsed for now
}
}
