% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{remove_clustered_variants}
\alias{remove_clustered_variants}
\title{Remove clustered sequence variants}
\usage{
remove_clustered_variants(
  se,
  txdb,
  regions = c("transcript", "genome"),
  variant_dist = 100
)
}
\arguments{
\item{se}{SummarizedExperiment containing editing sites}

\item{txdb}{A TxDb object}

\item{regions}{One of "transcript" and/or "genome", coordinate space to use
to examine distances between variants.}

\item{variant_dist}{distance in nucleotides for determining clustered variants}
}
\description{
Sequence variants of multiple allele types (e.g. A->G, C->T)
in nearby regions can be due to misalignments. This function will remove
variants if multiple allele types are present within a given distance
in genomic coordinate space or transcriptome coordinate space.
}
