% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{remove_splice_variants}
\alias{remove_splice_variants}
\title{Remove sites near splice sites}
\usage{
remove_splice_variants(se, txdb, splice_site_dist = 4, ignore.strand = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment with editing sites}

\item{txdb}{A TxDb object}

\item{splice_site_dist}{distance to splice site}

\item{ignore.strand}{if TRUE do not consider strand when comparing editing sites to
splice sites}
}
\description{
This function will remove editing sites found in regions proximal
to annotated splice junctions.
}
