% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse_adar_ifn}
\alias{rse_adar_ifn}
\title{Sample data for RNA editing analysis}
\format{
\subsection{\code{rse_adar_ifn}}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{class: RangedSummarizedExperiment
dim: 74 2
metadata(0):
assays(7): Var nRef ... nC nG
rownames(74): SSR3_102_- SSR3_125_- ... DHFR_430_- DHFR_513_-
rowData names(1): Ref
colnames(2): wt adar1_ko
colData names(1): sample
}\if{html}{\out{</div>}}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/bioproject/PRJNA386593}
}
\usage{
data(rse_adar_ifn)
}
\value{
RangedSummarizedExperiment populated with pileup data
}
\description{
A subset of data from an RNA-seq experiment to measure
the effects of IFN treatment of cell lines with wild-type
or ADAR1-KO
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/29395325/}
}
\keyword{datasets}
