% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-utils.R
\name{sc_editing}
\alias{sc_editing}
\title{Calculate editing frequencies}
\usage{
sc_editing(
  bamfn,
  fafn,
  bedfn,
  cell_barcodes,
  cell_bc_tag = "CB",
  min_counts = 25L,
  assay_cols = c("nA", "nG"),
  ...,
  BPPARAM = SerialParam(),
  verbose = TRUE
)
}
\arguments{
\item{bamfn}{BAM file name}

\item{fafn}{FASTA file name}

\item{bedfn}{BED file containing editing sites}

\item{cell_barcodes}{List of character vectors containing cell barcodes
to query. See examples for specification.}

\item{cell_bc_tag}{tag in bam file containing cell barcodes}

\item{min_counts}{Minimum read counts required to consider a site for editing.
This is calculated across all reads in the bamfile prior to running \code{get_pileup()}
per cell to remove low-frequency events. Read alignments are required to be
non-duplicate, primary, QC passing, and non-supplemental.}

\item{assay_cols}{assays to store in returned se. Set to "A" and "G". Note that
storing multiple assays can require large amounts of memory.}

\item{...}{additional arguments passed to \verb{[get_pileup()]}.}

\item{BPPARAM}{BiocParallel instance. Parallel computation occurs across
each entry in the cell_barcodes list}

\item{verbose}{Display messages}
}
\description{
Calculate editing frequencies
}
