% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.R
\name{annot_from_gr}
\alias{annot_from_gr}
\title{Annotate a RangedSummarizedExperiment using Granges objects}
\usage{
annot_from_gr(obj, gr, cols_to_map, ...)
}
\arguments{
\item{obj}{RangedSummarizedExperiment or GRanges object}

\item{gr}{GRanges with annotations to map to obj}

\item{cols_to_map}{character vector of columns from gr to map to obj. If the
vector has names, the names will be the column names in the output obj}

\item{...}{additional arguments to pass to \code{\link[GenomicRanges:findOverlaps-methods]{GenomicRanges::findOverlaps()}}}
}
\value{
Either a SummarizedExperiment or GRanges object with additional
annotations provided by the supplied GRanges object.
}
\description{
Utility function to map annotations from GRanges to rowData of
SummarizedExperiment or GRanges object. If multiple features overlap then
they will be concatenated as comma separated values.
}
\examples{
example(merge_pileups, echo = FALSE)

library(SummarizedExperiment)

gr <- GRanges(rep(c("SSR3", "SPCS3"), c(5, 15)),
  IRanges(seq(1, 500, by = 25), width = 50),
  strand = "+"
)

gr$feature <- sample(1:100, size = 20)
gr$id <- sample(LETTERS, size = 20)

se <- annot_from_gr(se, gr, c(feature_set = "feature", "id"))
rowData(se)

}
