% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{annot_snps}
\alias{annot_snps}
\title{Annotate snps}
\usage{
annot_snps(obj, ...)
}
\arguments{
\item{obj}{GRanges or SummarizedExperiment  object}

\item{dbsnp}{SNPlocs package, see available packages from
\code{\link[BSgenome:injectSNPs]{BSgenome::available.SNPs()}}}

\item{chrom}{only operate on a specified chromosome}

\item{col_to_aggr}{column from SNPlocs package to add to
input. If multiple SNPs overlap these values will be concatenated
as comma separated values.}

\item{drop}{If TRUE, remove sites overlap SNPs}
}
\value{
Either a GRanges or SummarizedExperiment object with
a new column "snp" added with information from "col_to_aggr"
}
\description{
Annotate snps
}
