% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/annot.R
\name{annot_snps}
\alias{annot_snps}
\alias{annot_snps.GRanges}
\alias{annot_snps.SummarizedExperiment}
\title{Annotate known SNP positions}
\usage{
annot_snps(obj, ...)

\method{annot_snps}{GRanges}(
  obj,
  dbsnp,
  chrom = NULL,
  col_to_aggr = "RefSNP_id",
  drop = FALSE,
  ...
)

\method{annot_snps}{SummarizedExperiment}(
  obj,
  dbsnp,
  chrom = NULL,
  col_to_aggr = "RefSNP_id",
  drop = FALSE,
  ...
)
}
\arguments{
\item{obj}{GRanges or SummarizedExperiment  object}

\item{...}{For the generic, further arguments to pass to specific methods.
Unused for now.}

\item{dbsnp}{SNPlocs package, see available packages from
\code{\link[BSgenome:injectSNPs]{BSgenome::available.SNPs()}}}

\item{chrom}{only operate on a specified chromosome}

\item{col_to_aggr}{column from SNPlocs package to add to
input. If multiple SNPs overlap these values will be concatenated
as comma separated values.}

\item{drop}{If TRUE, remove sites overlap SNPs}
}
\value{
Either a GRanges or SummarizedExperiment object with
a new column "snp" added with information from "col_to_aggr"
}
\description{
This function will annotate a GRanges or rowRanges of
a SummarizedExperiment with SNP positions from a SNP package
}
\examples{
\dontrun{
if (require(SNPlocs.Hsapiens.dbSNP144.GRCh38)) {
  gr <- GRanges(rep("22", 10),
                IRanges(seq(10510077,
                            10610077,
                            by = 1000)[1:10],
                        width = 250),
                strand = "+")
  annot_snps(gr, SNPlocs.Hsapiens.dbSNP144.GRCh38)
}
}
}
\seealso{
\href{https://bioconductor.org/packages/release/data/annotation/html/SNPlocs.Hsapiens.dbSNP144.GRCh38.html}{SNPlocs.Hsapiens.dbSNP144.GRCh38}
}
