% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-utils.R
\name{build_tag_index}
\alias{build_tag_index}
\title{Build index for tag sorted bam file}
\usage{
build_tag_index(
  bamfile,
  tag = "CB",
  n_records_to_check = 1e+06,
  overwrite = TRUE
)
}
\arguments{
\item{bamfile}{tag sorted bamfile}

\item{tag}{name of tag in bamfile used for sorting. The tag must be of type
"Z".}

\item{n_records_to_check}{The number of bam records to query to validate that
the tag is present and of the correct type. Set to 0 to disable checks.}

\item{overwrite}{if TRUE, regenerate index if it already exists}
}
\value{
name of index generated, which is the bam file + ".bri"
}
\description{
Build index for tag sorted bam file
}
\examples{
bam_fn <- system.file("extdata", "5k_neuron_mouse_xf25_1pct_cbsort.bam", package = "raer")
build_tag_index(bam_fn)

bam_fn <- system.file("extdata", "5k_neuron_mouse_xf25_1pct_ubsort.bam", package = "raer")
build_tag_index(bam_fn, tag = "UB")

}
