% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AEI.R
\name{calc_AEI}
\alias{calc_AEI}
\title{Calculate the Adenosine Editing Index (AEI)}
\usage{
calc_AEI(
  bam_fn,
  fasta_fn,
  alu_ranges = NULL,
  txdb = NULL,
  snp_db = NULL,
  BPPARAM = SerialParam(),
  verbose = FALSE
)
}
\arguments{
\item{bam_fn}{bam file}

\item{fasta_fn}{fasta}

\item{alu_ranges}{GRanges with regions to query for calculating the AEI,
typically ALU repeats.}

\item{txdb}{A txdb object, if supplied,  will be used to subset the alu_ranges to
those found overlapping genes.}

\item{snp_db}{a SNP package, if supplied, will be used to exclude polymorphic positions
prior to calculating the AEI.}

\item{BPPARAM}{A \link{BiocParallelParam} object for specifying parallel options for
operating over chromosomes.}
}
\value{
A named list with the AEI index computed for all allelic combinations.
If correctly computed the signal from the A_G index should be higher than other
alleles (T_C), which are most likely derived from noise or polymorphisms.
}
\description{
The Adenosine Editing Index describes the magnitude of A-to-I editing
in a sample. The index is a weighted average of editing events (G bases) observed
at A positions. The vast majority A-to-I editing occurs in ALU elements in the human
genome, and these regions have a high A-to-I editing signal compared to other regions
such as coding exons. This function will perform pileup at specified repeat regions and
return a summary AEI metric.
}
\examples{
bamfn <- system.file("extdata", "SRR5564277_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")
calc_AEI(bamfn, fafn)

}
\references{
Roth, S.H., Levanon, E.Y. & Eisenberg, E. Genome-wide quantification of ADAR adenosine-to-inosine RNA editing activity. Nat Methods 16, 1131–1138 (2019). https://doi.org/10.1038/s41592-019-0610-9
}
