% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_editing.R
\name{calc_edit_frequency}
\alias{calc_edit_frequency}
\title{Adds editing frequencies}
\usage{
calc_edit_frequency(
  se,
  edit_from = NULL,
  edit_to = NULL,
  drop = FALSE,
  replace_na = TRUE,
  edit_frequency = 0,
  min_count = 1
)
}
\arguments{
\item{se}{A SummarizedExperiment object created by \code{merge_pileups}}

\item{edit_from}{This should be a nucleotide (A, C, G, or T)
corresponding to the nucleotide you expect in the reference. Ex. for A to I
editing events, this would be "A". If NULL, then editing frequencies will be
calculated using the \code{nVar} and \code{nRef} values.}

\item{edit_to}{This should be a nucleotide (A, C, G, or T) and should
correspond to the nucleotide you expect after the editing event. Ex. for A
to I editing events, this would be "G". If NULL, then editing frequencies
will be calculated using the \code{nVar} and \code{nRef} values.}

\item{drop}{If TRUE, the summarizedExperiment returned will only retain sites
matching the specified \code{edit_from} and \code{edit_to} bases.}

\item{replace_na}{If TRUE, NA and NaN editing frequencies will be coerced to
0.}

\item{edit_frequency}{The edit frequency cutoff used when calculating the
number of sites. Set to 0 to require any non-zero editing frequency. The
number of sites is stored as n_sites in the colData.}

\item{min_count}{The minimum number of reads required when enumerating number
of editing sites detected.}
}
\value{
SummarizedExperiment supplemented with \code{edit_freq} assay.
}
\description{
Adds editing frequencies to an existing
SummarizedExperiment object (created by merge_pileups\verb{). The SummarizedExperiment with a new assay for editing frequencies for each site (}edit_freq\verb{), depth of coverage computed using the indicatededited nucleotides (}depth`) and new colData
columns with the number of edited sites (n_sites) and the
fraction of edits (edit_idx) is returned.
}
\examples{
library(SummarizedExperiment)
se <- calc_edit_frequency(rse_adar_ifn)
assay(se, "edit_freq")[1:5, ]

}
