% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_scAEI.R
\name{calc_scAEI}
\alias{calc_scAEI}
\alias{get_scAEI_sites}
\title{Calculate the Adenosine Editing Index (AEI) in single cells}
\usage{
calc_scAEI(
  bamfiles,
  sites,
  cell_barcodes,
  param = FilterParam(),
  edit_from = "A",
  edit_to = "G",
  output_dir = NULL,
  return_sce = FALSE,
  ...
)

get_scAEI_sites(fasta_fn, genes, alus, edit_from = "A", edit_to = "G")
}
\arguments{
\item{bamfiles}{a path to a BAM file (for 10x libraries), or a vector of paths
to BAM files (smart-seq2). Can be supplied as a character vector, \link{BamFile}, or
\link{BamFileList}.}

\item{sites}{a GRanges object produced by \link{get_scAEI_sites} containing sites to process.}

\item{cell_barcodes}{A character vector of single cell barcodes to process. If
processing multiple BAM files (e.g. smart-seq-2), provide a character vector
of unique identifiers for each input BAM, to name each BAM file in the output files.}

\item{param}{object of class \code{\link[=FilterParam]{FilterParam()}} which specify various filters to
apply to reads and sites during pileup.}

\item{edit_from}{This should correspond to the base
(\code{A}, \code{C}, \code{G}, \code{T}) you expect in the reference. Ex. for A to I
editing events, this would be \code{A}.}

\item{edit_to}{This should correspond to the base
(\code{A}, \code{C}, \code{G}, \code{T}) you expect in an edited site. Ex. for A
to I editing events, this would be \code{G}.}

\item{return_sce}{if \code{TRUE}, data is returned as a SingleCellExperiment, if
\code{FALSE} a \link{DataFrame} containing computed AEI values will be returned.}

\item{...}{additional arguments to \link{pileup_cells}}

\item{fasta_fn}{a path to a BAM file (for 10x libraries), or a vector of paths
to BAM files (smart-seq2). Can be supplied as a character vector, \link{BamFile}, or
\link{BamFileList}.}

\item{genes}{A \link{GRanges} object with gene coordinates.Alternatively a \link{TxDb} object,
which if supplied, will be used extract gene coordinates.}

\item{alus}{\link{GRanges} with repeat regions to query for calculating the AEI,
typically ALU repeats. The strand of the supplied intervals will be ignored for
defining repeat regions.}

\item{output_directory}{Output directory for \code{nRef} and \code{nAlt} sparseMatrix files.
If NULL, a temporary directory will be used.}
}
\value{
A \code{SingleCellExperiment} or a \code{DataFrame} containing computed AEI values,
count of editing events (n_alt), and count of referenc events (n_ref) per cell.
}
\description{
The Adenosine Editing Index describes the magnitude of A-to-I
editing in a sample. The index is a weighted average of editing events (G
bases) observed at A positions. The vast majority A-to-I editing occurs in
ALU elements in the human genome, and these regions have a high A-to-I
editing signal compared to other regions such as coding exons. This
function will perform pileup at specified repeat regions and return a
summary AEI metric.

The Adenosine Editing Index describes the magnitude of A-to-I
editing in a sample. The index is a weighted average of editing events (G
bases) observed at A positions. The vast majority A-to-I editing occurs in
ALU elements in the human genome, and these regions have a high A-to-I
editing signal compared to other regions such as coding exons. This
function will perform pileup at specified repeat regions and return a
summary AEI metric.
}
\references{
Roth, S.H., Levanon, E.Y. & Eisenberg, E. Genome-wide
quantification of ADAR adenosine-to-inosine RNA editing activity. Nat Methods
16, 1131–1138 (2019). https://doi.org/10.1038/s41592-019-0610-9

Roth, S.H., Levanon, E.Y. & Eisenberg, E. Genome-wide
quantification of ADAR adenosine-to-inosine RNA editing activity. Nat Methods
16, 1131–1138 (2019). https://doi.org/10.1038/s41592-019-0610-9
}
