% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BedFile.R
\name{close.BedFile}
\alias{close.BedFile}
\title{Close a BedFile index connection}
\usage{
\method{close}{BedFile}(con, ...)
}
\arguments{
\item{con}{BedFile class}

\item{...}{present for consistency
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
bed <- indexBed(bed_fn)
close(bed)}
}
\description{
Close a BedFile index connection
}
