% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_se.R
\name{create_se}
\alias{create_se}
\title{Create RangedSummarizedExperiment}
\usage{
create_se(
  plps,
  rowdata_cols = c("Ref"),
  assay_cols = c("Var", "nRef", "nVar", "nA", "nT", "nC", "nG"),
  sample_names = NULL
)
}
\arguments{
\item{plps}{results from running get_pileup.R, can be one result, a list
of results, or a named list of results. If a named list is give, the colData
will be named using the names in the list.}

\item{rowdata_cols}{character vector of columns to store in rowData
. Values must be the same across all assays (excluding NA).}

\item{assay_cols}{character vector of columns to store as assays}

\item{sample_names}{OPTIONAL A list of names to be added to the SE object.
If no sample names are given and pileup_res is not a named list, then
default names (ie sample_1, sample_2, ..., sample_n) will be given and
a warning will be printed.}
}
\description{
This function will take either a single result from running pileup_res or a list of
results (ie for different samples) from running pileup_res and will return a summarized
experiment object that contains assays for each column in the pileup_res output.
Currently this doesn't add any sample info and can only take a list of results.
Potential added functionality would be to allow the user to give an existing
SE object as input or to give sample meta data.
}
\examples{
library(SummarizedExperiment)
bamfn <- system.file("extdata", "SRR5564269_Aligned.sortedByCoord.out.md.bam", package = "raer")
bam2fn <- system.file("extdata", "SRR5564277_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")

plps <- get_pileup(c(bamfn, bam2fn), fafn)
names(plps) <- c("sample1", "sample2")
se <- create_se(plps)

assays(se)

colData(se)

rowRanges(se)

}
