% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{download_GSE99249}
\alias{download_GSE99249}
\title{Download GSE99249 BAM files and related data}
\usage{
download_GSE99249(path)
}
\arguments{
\item{path}{path to directory to download data}
}
\value{
A named list with paths to BAM files, a FASTA file and a bed
file of known editing sites from hg38 chromosome 18.
}
\description{
This function will download ~ 250 MB of data.
}
