% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-utils.R
\name{filter_by_coverage}
\alias{filter_by_coverage}
\title{Identify sites with read coverage}
\usage{
filter_by_coverage(bamfile, gr, min_counts, param = NULL, verbose = FALSE, ...)
}
\arguments{
\item{bamfile}{Bam file name}

\item{gr}{GRanges, or file name of BED file, containing site to
examine}

\item{min_counts}{minimum reads counts to require}

\item{param}{A ScanBamParam object specifying how reads should be filtered.
If not supplied the default behavior will ignore exclude alignments marked
as secondary, duplicate, supplementary, or QC-fail.}

\item{verbose}{If TRUE, print messages}

\item{...}{Additional arguments to supply to \code{\link[GenomicAlignments:coverage-methods]{GenomicAlignments::coverage()}}}
}
\description{
This function will compute read coverage at sites
listed in either a GRanges or BED file. Sites passing with at least
\code{min_counts} read depth will be returned. This function will
compute coverage across all alignments and positions,
ignoring any cell-barcodes. Generally it is faster to compute coverage
across all positions, then filter to those in the GRanges object, rather
than only querying the supplied sites.
}
