% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{filter_clustered_variants}
\alias{filter_clustered_variants}
\title{Filter out clustered sequence variants}
\usage{
filter_clustered_variants(
  se,
  txdb,
  regions = c("transcript", "genome"),
  variant_dist = 100
)
}
\arguments{
\item{se}{\code{SummarizedExperiment::SummarizedExperiment} containing editing sites}

\item{txdb}{\code{GenomicFeatures::TxDb}}

\item{regions}{One of \code{transcript} or \code{genome}, specifying the coordinate
system for calculating distances between variants.}

\item{variant_dist}{distance in nucleotides for determining clustered
variants}
}
\value{
\code{SummarizedExperiment::SummarizedExperiment} with sites removed from
object dependent on filtering applied.
}
\description{
Sequence variants of multiple allele types (e.g., \code{A>G}, \code{A>C})
in nearby regions can be due to mis-alignment. Remove variants if multiple
allele types are present within a given distance in genomic or
transcriptome coordinate space.
}
\examples{
if (require(TxDb.Hsapiens.UCSC.hg38.knownGene)) {
  library(SummarizedExperiment)
  nrows <- 5
  ncols <- 6
  counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
  rowRanges <- GRanges(rep("chr1", 5),
    IRanges(c(12055, 12174, 12194, 12719, 12889), width = 1),
    strand = rep("+", 5)
  )
  mcols(rowRanges)$Var <- c("AG", "AT", "AC", "TC", "GC")
  colData <- DataFrame(
    Treatment = rep(c("adar_wt", "adar_ko"), 3),
    row.names = LETTERS[1:6]
  )
  rse <- SummarizedExperiment(
    assays = SimpleList(counts = counts),
    rowRanges = rowRanges, colData = colData
  )

  se <- remove_clustered_variants(rse, TxDb.Hsapiens.UCSC.hg38.knownGene)
  se
}

}
\seealso{
Other se-filters: 
\code{\link{filter_multiallelic}()},
\code{\link{filter_splice_variants}()}
}
\concept{se-filters}
