% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{filter_clustered_variants}
\alias{filter_clustered_variants}
\title{Filter out clustered sequence variants}
\usage{
filter_clustered_variants(
  se,
  txdb,
  regions = c("transcript", "genome"),
  variant_dist = 100
)
}
\arguments{
\item{se}{\code{SummarizedExperiment::SummarizedExperiment} containing editing sites}

\item{txdb}{\code{GenomicFeatures::TxDb}}

\item{regions}{One of \code{transcript} or \code{genome}, specifying the coordinate
system for calculating distances between variants.}

\item{variant_dist}{distance in nucleotides for determining clustered
variants}
}
\value{
\code{SummarizedExperiment::SummarizedExperiment} with sites removed from
object dependent on filtering applied.
}
\description{
Sequence variants of multiple allele types (e.g., \code{A>G}, \code{A>C})
in nearby regions can be due to mis-alignment. Remove variants if multiple
allele types are present within a given distance in genomic or
transcriptome coordinate space.
}
\examples{
if (require(TxDb.Hsapiens.UCSC.hg38.knownGene)) {
  rse <- filter_multiallelic(rse_adar_ifn)

  filter_clustered_variants(
    rse,
    TxDb.Hsapiens.UCSC.hg38.knownGene
  )
}

}
\seealso{
Other se-filters: 
\code{\link{filter_multiallelic}()},
\code{\link{filter_splice_variants}()}
}
\concept{se-filters}
