% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{filter_splice_variants}
\alias{filter_splice_variants}
\title{Filter out sites near splice sites}
\usage{
filter_splice_variants(se, txdb, splice_site_dist = 4, ignore.strand = FALSE)
}
\arguments{
\item{se}{\code{SummarizedExperiment::SummarizedExperiment} with editing sites}

\item{txdb}{\code{GenomicFeatures::TxDb}}

\item{splice_site_dist}{distance to splice site}

\item{ignore.strand}{if \code{TRUE}, ignore strand when comparing editing sites to
splice sites}
}
\description{
Remove editing sites found in regions proximal to annotated
splice junctions.
}
\examples{
if (require(TxDb.Hsapiens.UCSC.hg38.knownGene)) {
  library(SummarizedExperiment)
  nrows <- 5
  ncols <- 6
  counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
  rowRanges <- GRanges(rep("chr1", 5),
    IRanges(c(12055, 12174, 12194, 12719, 12889), width = 1),
    strand = rep("+", 5)
  )
  colData <- DataFrame(
    Treatment = rep(c("adar_wt", "adar_ko"), 3),
    row.names = LETTERS[1:6]
  )
  rse <- SummarizedExperiment(
    assays = SimpleList(counts = counts),
    rowRanges = rowRanges, colData = colData
  )

  se <- filter_splice_variants(rse, TxDb.Hsapiens.UCSC.hg38.knownGene)
  se
}

}
\seealso{
Other se-filters: 
\code{\link{filter_clustered_variants}()},
\code{\link{filter_multiallelic}()}
}
\concept{se-filters}
