% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{filter_splice_variants}
\alias{filter_splice_variants}
\title{Filter out sites near splice sites}
\usage{
filter_splice_variants(se, txdb, splice_site_dist = 4, ignore.strand = FALSE)
}
\arguments{
\item{se}{\code{SummarizedExperiment::SummarizedExperiment} with editing sites}

\item{txdb}{\code{GenomicFeatures::TxDb}}

\item{splice_site_dist}{distance to splice site}

\item{ignore.strand}{if \code{TRUE}, ignore strand when comparing editing sites to
splice sites}
}
\description{
Remove editing sites found in regions proximal to annotated
splice junctions.
}
\examples{
if (require(TxDb.Hsapiens.UCSC.hg38.knownGene)) {
  filter_splice_variants(
    rse_adar_ifn,
    TxDb.Hsapiens.UCSC.hg38.knownGene
  )
}

}
\seealso{
Other se-filters: 
\code{\link{filter_clustered_variants}()},
\code{\link{filter_multiallelic}()}
}
\concept{se-filters}
