% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup.R
\name{get_pileup}
\alias{get_pileup}
\alias{FilterParam}
\title{Generate base counts using pileup}
\usage{
get_pileup(
  bamfiles,
  fafile,
  bedfile = NULL,
  region = NULL,
  chroms = NULL,
  filterParam = FilterParam(),
  outfile_prefix = NULL,
  bedidx = NULL,
  bam_flags = NULL,
  reads = NULL,
  return_data = TRUE,
  BPPARAM = SerialParam(),
  bad_reads = NULL,
  verbose = FALSE
)

FilterParam(
  max_depth = 10000,
  min_base_quality = 20L,
  min_mapq = 0L,
  min_nucleotide_depth = 1L,
  library_type = "fr-first-strand",
  only_keep_variants = FALSE,
  trim_5p = 0L,
  trim_3p = 0L,
  indel_dist = 0L,
  splice_dist = 0L,
  homopolymer_len = 0L,
  max_mismatch_type = c(0L, 0L),
  min_read_qual = c(0L, 0L),
  ignore_query_Ns = FALSE
)
}
\arguments{
\item{bamfiles}{paths to 1 or more bam files}

\item{fafile}{path to fasta file}

\item{bedfile}{path to bed file with sites or regions to query}

\item{region}{samtools region query string (i.e. chr1:100-1000)}

\item{chroms}{chromosomes to process, not to be used with region}

\item{filterParam}{object of class \code{\link[=FilterParam]{FilterParam()}} which specify various
filters to apply to reads and sites during pileup.}

\item{outfile_prefix}{output prefix for tabix indexed files}

\item{bedidx}{A BedFile object, if supplied, pileup will use index
generated by \code{\link[=indexBed]{indexBed()}}}

\item{bam_flags}{bam flags to filter or keep, use \code{\link[Rsamtools:ScanBamParam-class]{Rsamtools::scanBamFlag()}}
to generate}

\item{reads}{if supplied a fasta file will be written with reads that pass filters
and contain variants}

\item{return_data}{return data as a Granges table?}

\item{BPPARAM}{A \link{BiocParallel} class to control parallel execution. Parallel
processing occurs per chromosome, so is disabled when run on a single region.}

\item{bad_reads}{a textfile containing read names to exclude from pileup. Readnames
should be formated as readid_1 or readid_2 or readid for paired end first read
paired-end second read or single end data.}

\item{verbose}{if TRUE, then report progress.}

\item{max_depth}{maximum read depth considered at each site}

\item{min_base_quality}{min base quality score to consider read for pileup}

\item{min_mapq}{minimum required MAPQ score, can be a vector of values
for each bam file}

\item{min_nucleotide_depth}{min read depth needed to report site}

\item{library_type}{read orientation, one of fr-first-strand,
fr-second-strand, unstranded, and genomic-unstranded. Can supply as a vector to specify for each
input bam. Unstranded library type will be reported based on read alignment.
genomic-unstranded will report all variants w.r.t the + strand.}

\item{only_keep_variants}{if TRUE, then only variant sites will be reported
(FALSE by default), can be a vector for each input bamfile}

\item{trim_5p}{Bases to trim from 5' ends of read alignments}

\item{trim_3p}{Bases to trim from 3' ends of read alignments}

\item{indel_dist}{Exclude read if site occurs within given
distance from indel event in the read}

\item{splice_dist}{Exclude read if site occurs within given
distance from splicing event in the read}

\item{homopolymer_len}{Exclude site if occurs within homopolymer of given
length}

\item{max_mismatch_type}{Exclude read if it has N different mismatch types
(e.g A-to-G, G-to-C, C-to-G, is 3 mismatch types) or Y # of mismatches,
must be supplied as a integer vector of length 2. e.g.
c(X, Y).}

\item{min_read_qual}{ignored for now}

\item{ignore_query_Ns}{ignored for now}
}
\value{
A list containing a \code{GRanges} object for each input bam file, or a vector
of the output tabixed file names if \code{return_data} is FALSE.
}
\description{
Generate base counts using pileup
}
\details{
Multiple bam files can be processed together, with files being
written for each bam file. In this mode the output regions will be consistent across all files.
The min_mapq, only_keep_variants, and library_type parameters can be specified for each input files.
}
\examples{
bamfn <- system.file("extdata", "SRR5564269_Aligned.sortedByCoord.out.md.bam", package = "raer")
bam2fn <- system.file("extdata", "SRR5564277_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")

plp <- get_pileup(bamfn, fafn)
plps <- get_pileup(c(bamfn, bam2fn), fafn)
fp <-FilterParam(only_keep_variants = TRUE,  min_nucleotide_depth = 55)
get_pileup(bamfn, fafn, filterParam = fp)

}
