% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pileup.R
\name{get_pileup}
\alias{get_pileup}
\title{Generate pileup}
\usage{
get_pileup(
  bamfile,
  fafile,
  bedfile = NULL,
  region = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL,
  min_reads = c(1L),
  min_base_qual = 20L,
  min_mapq = 0L,
  max_depth = 10000,
  library_type = "fr-first-strand",
  outfile = NULL,
  bedidx = NULL,
  required_flags = "0",
  filter_flags = "0",
  n_align = 1,
  n_align_tag = "NH",
  event_filters = c(0, 0, 0, 0),
  return_data = TRUE
)
}
\arguments{
\item{bamfile}{path to bam file}

\item{fafile}{path to fasta file}

\item{bedfile}{path to bed file with sitesu or regions to query}

\item{region}{samtools region query string (i.e. chr1:100-1000)}

\item{chrom}{chrom to process, not to be used with region}

\item{start}{start position, not to be used with region}

\item{end}{end position, not to be used with region}

\item{min_reads}{min read depth needed to report site}

\item{min_base_qual}{min base quality score to consider read for pileup}

\item{min_mapq}{minimum required MAPQ score}

\item{max_depth}{maximum read depth considered at each site}

\item{library_type}{read orientation, one of fr-first-strand,
fr-second-strand, or unstranded.}

\item{outfile}{output tabix indexed file?}

\item{bedidx}{A BedFile object, if supplied, pileup will use index
generated by \code{\link[=indexBed]{indexBed()}}.}

\item{required_flags}{bam tags required, pass as integer flag value}

\item{filter_flags}{bam tags to filter, pass as integer flag value}

\item{event_filters}{integer vector of length 4 with values indicating
distance from pileup position to features to filter. Distance to trim from
5' ends of reads, distance from splicing events, distance from indels, and
length of homopolymer stretch used to determine if a site is in a homopolymer.
default is no filters applied (e.g. c(0,0,0,0)).}

\item{return_data}{return data as a Granges table?}
}
\description{
Generate pileup
}
