% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pileup.R
\name{get_pileup}
\alias{get_pileup}
\title{Generate base counts using pileup}
\usage{
get_pileup(
  bamfile,
  fafile,
  bedfile = NULL,
  region = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL,
  min_reads = c(1L),
  min_base_qual = 20L,
  min_mapq = c(0L),
  max_depth = 10000,
  library_type = "fr-first-strand",
  outfile = NULL,
  bedidx = NULL,
  bam_flags = NULL,
  n_align = 0,
  n_align_tag = "NH",
  event_filters = c(0, 0, 0, 0),
  only_keep_variants = FALSE,
  outbam = NULL,
  return_data = TRUE
)
}
\arguments{
\item{bamfile}{paths to 1 or 2 bam files, pass 2 files to}

\item{fafile}{path to fasta file}

\item{bedfile}{path to bed file with sites or regions to query}

\item{region}{samtools region query string (i.e. chr1:100-1000)}

\item{chrom}{chrom to process, not to be used with region}

\item{start}{start position, not to be used with region}

\item{end}{end position, not to be used with region}

\item{min_reads}{min read depth needed to report site}

\item{min_base_qual}{min base quality score to consider read for pileup}

\item{min_mapq}{minimum required MAPQ score}

\item{max_depth}{maximum read depth considered at each site}

\item{library_type}{read orientation, one of fr-first-strand,
fr-second-strand, or unstranded.}

\item{outfile}{output tabix indexed file?}

\item{bedidx}{A BedFile object, if supplied, pileup will use index
generated by \code{\link[=indexBed]{indexBed()}}}

\item{bam_flags}{bam flags to filter or keep, use \code{\link[Rsamtools:ScanBamParam-class]{Rsamtools::scanBamFlag()}}
to set}

\item{n_align}{of valid alignments to use for each read, based on tag value

supplied to n_align_tag. NULL disables filtering by n_align_tag
value}

\item{n_align_tag}{Tag name that containing # of valid alignments}

\item{event_filters}{integer vector of length 4 with values indicating
distance from pileup position to features to filter. Distance to trim from
5' ends of reads, distance from splicing events, distance from indels, and
length of homopolymer stretch used to determine if a site is in a homopolymer.
default is no filters applied (e.g. c(0,0,0,0)).}

\item{only_keep_variants}{if TRUE, then only variant sites will be reported
(FALSE by default)}

\item{outbam}{if supplied a bam file will be written with reads that pass filters
and contain variants}

\item{return_data}{return data as a Granges table?}
}
\description{
Generate base counts using pileup
}
\details{
To compare RNA-seq to DNA-seq 2 bam files can be supplied to the \code{bamfile}
argument, with the first corresponding to RNA-seq and second DNA-seq. The \code{min_reads}
and \code{min_mapq} arguments can accept 2 values (as numeric vector) corresponding to filters
for RNA and DNA-seq respectively. The \code{min_base_qual}, \code{n_align},
\code{n_align_tag}, and \code{event_filters} params will be applied to both files.

The library type argument will only apply to the 1st bam file. The second will be set to
"unstranded"
}
