% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-BedFile.R
\docType{class}
\name{BedFile-class}
\alias{BedFile-class}
\alias{.BedFile}
\alias{indexBed}
\alias{indexBed,character-method}
\alias{close.BedFile}
\title{A reference class for generating and storing index of BedFile}
\usage{
indexBed(file, ...)

\S4method{indexBed}{character}(file)

\method{close}{BedFile}(con, ...)
}
\arguments{
\item{file}{path to bed file}

\item{...}{present for consistency}

\item{con}{BedFile class}
}
\value{
Return an object of class BedFile which contains a pointer
to the bed file index.
}
\description{
Class to store information about a bedfile used
for indicating regions for pileup
}
\section{Fields}{

\describe{
\item{\code{.extptr}}{an externalptr to c-level samtools bedindex}

\item{\code{path}}{filepath to bed file}

\item{\code{open}}{logical indicating if the index is open}
}}


\examples{
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
indexBed(bed_fn)
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
bed <- indexBed(bed_fn)
close(bed)
}
