% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BedFile.R
\name{indexBed}
\alias{indexBed}
\alias{indexBed,character-method}
\title{Generate a samtools style index for a BedFile}
\usage{
indexBed(file, ...)

\S4method{indexBed}{character}(file)
}
\arguments{
\item{file}{path to bed file}

\item{...}{For the generic, further arguments to pass to specific methods.
Unused for now.}
}
\value{
Return an object of class BedFile which contains a pointer
to the bed file index.
}
\description{
Generate a samtools style index for a BedFile
}
\examples{
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
indexBed(bed_fn)
}
