% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{perform_de}
\alias{perform_de}
\title{Perform differential editing
Uses either edgeR or DESeq2 to perform differential editing analysis.
This will work for simple designs that have 1 treatment and 1 control.
For more complex designs, we suggest you perform your own.}
\usage{
perform_de(
  deobj,
  type = "edgeR",
  sample_col = "sample",
  condition_col = "condition",
  condition_control = NULL,
  condition_treatment = NULL
)
}
\arguments{
\item{deobj}{A SummarizedExperiment object prepared for de by \code{prep_for_de}}

\item{type}{OPTIONAL if edgeR or DESeq should be run. Default is edgeR}

\item{condition_col}{OPTIONAL the name of the column from colData(deobj) that
contains your treatment information. Default is condition, If you do not have a
column named "condition", you must provide the appropriate condition column}

\item{condition_control}{The name of the control condition. This must be a variable
in your condition_col of colData(deobj). No default provided.}

\item{condition_treatment}{The name of the treatment condition. This must be a variable
in your condition_col of colData(deobj).}

\item{sample_column}{OPTIONAL the name of the column from colData(deobj) that
contains your sample information. Default is sample. If you do not have a
column named "sample", you must provide the appropriate sample column}
}
\description{
At the moment, this function will only find editing events specific
to the treatment, but it will be pretty straight forward to add other
possible return values.
}
