% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-pileup.R
\name{pileup_cells}
\alias{pileup_cells}
\title{Pileup sites per cell}
\usage{
pileup_cells(
  bamfile,
  sites,
  cell_barcodes,
  output_directory,
  bam_flags = NULL,
  chroms = NULL,
  umi_tag = "UB",
  cb_tag = "CB",
  fp = FilterParam(),
  BPPARAM = SerialParam(),
  return_sce = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{bamfile}{BAM file name}

\item{sites}{a GRanges object containing sites to process. See examples for
valid formatting.}

\item{cell_barcodes}{A character vector of single cell barcodes to process.}

\item{output_directory}{Output directory for output files. Will be generated
if it doesn't exist.}

\item{bam_flags}{bam flags to filter or keep, use \code{\link[Rsamtools:ScanBamParam-class]{Rsamtools::scanBamFlag()}}
to generate}

\item{chroms}{A character vector of chromosomes to process, if supplied,
only sites present in the listed chromosomes will be processed}

\item{umi_tag}{tag in bam containing the UMI sequence}

\item{cb_tag}{tag in bam containing the cell barcode sequence}

\item{BPPARAM}{BiocParallel instance. Parallel computation occurs across
chromosomes.}

\item{return_sce}{if \code{TRUE}, data is returned as a SingleCellExperiment,
if \code{FALSE} a character vector of the output files, specified by \code{outfile_prefix},
will be returned.}

\item{verbose}{Display messages}

\item{filterParam}{object of class \code{\link[=FilterParam]{FilterParam()}} which specify various
filters to apply to reads and sites during pileup. Note that the \code{min_variant_reads}
parameter, if set > 0, specifies the number of variant reads at a site required in
order to report a site. E.g. if set to 2, then at least 2 reads (from any cell) must
have a variant in order to report the site. The default of 0 reports all sites present
in the \code{sites} object.}
}
\value{
Returns either a SingleCellExperiment, or character vector of paths
to the files produced.
}
\description{
This function will perform a pileup operation at specified sites,
returning counts for Reference (e.g. unedited) or Alternate (e.g. editing) bases.
Current functionality will process a 10x genomic's style library, from a aligned
bam file containing a tag with a cell-barcode and a tag with a UMI.

The \code{sites} parameter specifies sites to pileup. This must be a GRanges object
with 1 base intervals, a strand (+ or -), and supplemented with metadata
columns named \code{ref} and \code{alt} containing the reference and alternate base to query. See
examples for an example format.

At each site, bases from overlapped reads will be examined, and counts of each ref
and alt base enumerated for each cell barcode present. A single base will be
counted once for each UMI sequence present in each cell.
}
