% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raer.R
\docType{package}
\name{raer}
\alias{raer}
\alias{raer-package}
\title{raer: RNA editing tools in R}
\description{
Toolkit for identification, visualization, and statistical testing of RNA editing signals from within R. Provides support for bulk-RNA-seq, single cell RNA-seq datasets, and provides generel methods for extraction of allelic read counts from alignment files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rnabioco.github.io/raer}
  \item \url{https://github.com/rnabioco/raer}
  \item Report bugs at \url{https://github.com/rnabioco/raer/issues}
}

}
\author{
\strong{Maintainer}: Kent Riemondy \email{kent.riemondy@gmail.com} (\href{https://orcid.org/0000-0003-0750-1273}{ORCID})

Authors:
\itemize{
  \item Kristen Wells-Wrasman \email{kristen.wells-wrasman@cuanschutz.edu} (\href{https://orcid.org/0000-0002-7466-8164}{ORCID})
}

Other contributors:
\itemize{
  \item Ryan Sheridan \email{ryan.sheridan@cuanschutz.edu} (\href{https://orcid.org/0000-0003-4012-3147}{ORCID}) [contributor]
  \item Jay Hesselberth \email{jay.hesselberth@gmail.com} (\href{https://orcid.org/0000-0002-6299-179X}{ORCID}) [contributor]
  \item RNA Bioscience Initiative [copyright holder, funder]
}

}
