% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raer.R
\docType{package}
\name{raer}
\alias{raer}
\alias{raer-package}
\title{raer: RNA editing tools in R}
\description{
Tools for identifying rna editing sites
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/rnabioco/raer}
  \item Report bugs at \url{https://github.com/rnabioco/raer/issues}
}

}
\author{
\strong{Maintainer}: Kent Riemondy \email{kent.riemondy@gmail.com}

}
