% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup.R
\name{read_pileup}
\alias{read_pileup}
\title{Read pileup, indexed by tabix}
\usage{
read_pileup(tbx_fn, region = NULL)
}
\arguments{
\item{tbx_fn}{filename}

\item{region}{region to read from file, samtools style
region specifiers are supported.}
}
\description{
Read pileup, indexed by tabix
}
\examples{
bamfn <- system.file("extdata", "SRR5564269_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")
plp_fn <- tempfile()
plp <- get_pileup(bamfn, fafn, return_data = FALSE, outfile_prefix = plp_fn)
read_pileup(plp)
unlink(c(plp, plp_fn, paste0(plp, ".tbi")))
}
