% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_tabix}
\alias{read_tabix}
\alias{get_tabix_chroms}
\title{Read in tabix indexed file as a data.frame}
\usage{
read_tabix(
  filename,
  region = ".",
  numeric_cols = c(2, 6:12),
  col_names = PILEUP_COLS
)

get_tabix_chroms(filename)
}
\arguments{
\item{filename}{path to indexed tabix file}

\item{region}{samtools region query string (i.e. chr1:100-1000)}

\item{numeric_cols}{columns to convert to numeric as a integer vector (one-based index)}

\item{col_names}{column names in the output data.frame}
}
\description{
Read in tabix indexed file as a data.frame

List chromosomes in a tabix index
}
\examples{
bamfn <- system.file("extdata", "SRR5564269_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")
plp_fn <- tempfile()
plp <- get_pileup(bamfn, fafn, return_data = FALSE, outfile_prefix = plp_fn)

head(read_tabix(plp))

read_tabix(plp, region = "SPCS3:498-500")

get_tabix_chroms(plp)

unlink(c(plp, plp_fn, paste0(plp, ".tbi")))
}
