% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{remove_clustered_variants}
\alias{remove_clustered_variants}
\title{Remove clustered sequence variants}
\usage{
remove_clustered_variants(
  se,
  txdb,
  regions = c("transcript", "genome"),
  variant_dist = 100
)
}
\arguments{
\item{se}{SummarizedExperiment containing editing sites}

\item{txdb}{A TxDb object}

\item{regions}{One of "transcript" and/or "genome", coordinate space to use
to examine distances between variants.}

\item{variant_dist}{distance in nucleotides for determining clustered variants}
}
\description{
Sequence variants of multiple allele types (e.g. A->G, C->T)
in nearby regions can be due to misalignments. This function will remove
variants if multiple allele types are present within a given distance
in genomic coordinate space or transcriptome coordinate space.
}
\examples{
if(require(TxDb.Hsapiens.UCSC.hg38.knownGene)){
  nrows <- 5; ncols <- 6
  counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
  rowRanges <- GRanges(rep("chr1", 5),
                    IRanges(c(12055, 12174, 12194, 12719, 12889), width=1),
                    strand=rep("+", 5))
  mcols(rowRanges)$Var <- c("AG", "AT", "AC", "TC", "GC")
  colData <- DataFrame(Treatment=rep(c("adar_wt", "adar_ko"), 3),
                       row.names=LETTERS[1:6])
  rse <- SummarizedExperiment(assays=SimpleList(counts=counts),
                           rowRanges=rowRanges, colData=colData)

  se <- remove_clustered_variants(rse, TxDb.Hsapiens.UCSC.hg38.knownGene)
  se
}
}
