% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_se.R
\name{remove_splice_variants}
\alias{remove_splice_variants}
\title{Remove sites near splice sites}
\usage{
remove_splice_variants(se, txdb, splice_site_dist = 4, ignore.strand = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment with editing sites}

\item{txdb}{A TxDb object}

\item{splice_site_dist}{distance to splice site}

\item{ignore.strand}{if TRUE do not consider strand when comparing editing sites to
splice sites}
}
\description{
This function will remove editing sites found in regions proximal
to annotated splice junctions.
}
\examples{
\dontrun{
if(require(TxDb.Hsapiens.UCSC.hg38.knownGene)){
  library(SummarizedExperiment)
  nrows <- 5; ncols <- 6
  counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
  rowRanges <- GRanges(rep("chr1", 5),
                    IRanges(c(12055, 12174, 12194, 12719, 12889), width=1),
                    strand=rep("+", 5))
  colData <- DataFrame(Treatment=rep(c("adar_wt", "adar_ko"), 3),
                       row.names=LETTERS[1:6])
  rse <- SummarizedExperiment(assays=SimpleList(counts=counts),
                           rowRanges=rowRanges, colData=colData)

  se <- remove_splice_variants(rse, TxDb.Hsapiens.UCSC.hg38.knownGene)
  se
}
}
}
